RAG/lab1/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.24
langchain-ollama==0.3.3
